skip service config if already present

+++ Deploy.sh
@@ -1,7 +1,7 @@
 #!/usr/bin/env bash
 
 echo "Filebeats install and create serviced entry"
-curl -L -O https://artifacts.elastic.co/downloads/beats/filebeat/filebeat-7.0.0-amd64.deb
+curl --silent -L -O https://artifacts.elastic.co/downloads/beats/filebeat/filebeat-7.0.0-amd64.deb
 sudo dpkg -i filebeat-7.0.0-amd64.deb
 
 #configure filebeats with propietary fields
@@ -12,13 +12,18 @@ sudo cp etc/haproxy.yml  /etc/filebeat/modules.d/haproxy.yml
 sudo filebeat modules enable haproxy
 
 #TODO create systemd service
-echo "Create systemd configuration "
-sudo cp systemd/filebeat.service /etc/systemd/system/filebeat.service
-#The unit file, source configuration file or drop-ins of filebeat.service changed on disk
-sudo systemctl daemon-reload
+echo "Create systemd configuration... "
+if [ ! -f /etc/systemd/system/filebeat.service ]; then
+    echo "filebeat service File not found!"
+	sudo cp systemd/filebeat.service /etc/systemd/system/filebeat.service
+	sudo systemctl daemon-reload
+	sudo systemctl enable filebeat
+else
+    echo "filebeat service File already present"
+fi
 
-sudo systemctl enable filebeat
+#The unit file, source configuration file or drop-ins of filebeat.service changed on disk
+#(re)start filebeat
 sudo systemctl start filebeat
 
-#restart filebeat
 sudo systemctl status filebeat
